@@ -2,7 +2,8 @@
 
 	<PropertyGroup>
 		<TargetFramework>netstandard2.0</TargetFramework>
-		<LangVersion>9.0</LangVersion>
+		<!--<TargetFramework>net7.0</TargetFramework>-->
+		<LangVersion>11.0</LangVersion>
 	</PropertyGroup>
 
 </Project>
