@@ -28,7 +28,7 @@
                                         }
                                         else
                                         {
-                                            @: Wipe @(i)
+                                            @: Wipe @(i) - @fightDetails.Duration
                                         }
                                     </MudText>
                                 </MudListItem>
@@ -41,7 +41,6 @@
         }
     </MudItem>
     <MudItem xs="8">
-
     </MudItem>
 </MudGrid>
 
