@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Diagnostics;
+using System.Diagnostics.Tracing;
 using System.Net.Http.Headers;
 using WoWCombatLogParser;
 
@@ -6,60 +8,29 @@ namespace WowCombatLogParser.App
 {
     internal class Program
     {
-        // WoWCombatLog-031423_155722.txt
+        static readonly string baseDirectory = @"E:\Games\World of Warcraft\_retail_\Logs";
 
         static async Task Main(string[] args)
         {
-            var log = args.FirstOrDefault();
-            if (!File.Exists(log))
-            {
-                Console.WriteLine($"Log file {log} not found.");
-                return;
-            }
-
-            var videoStart = DateTime.Parse($"{DateTime.Today:yyyy-MM-dd} 19:01:22").Add(TimeSpan.FromSeconds(15));
+            var log = Directory.GetFiles(baseDirectory, "WowCombatLog*.txt")
+                .Select(x => new FileInfo(x))
+                .OrderByDescending(x => x.LastAccessTime)
+                .Select(x => x.FullName)
+                .FirstOrDefault();
 
             var context = new ApplicationContext();
+
+            var sw = new Stopwatch();            
+            sw.Start();
+            
             var parser = context.CombatLogParser;
-
             parser.Filename = log;
-            var fights = parser.Scan(quickScan: true);            
-
-            string description = "";            
-            foreach (var boss in fights.GroupBy(x => x.Name))
-            {
-                int i = 0;
-                description += $"{boss.First().Name}\n{new string('=', 20)}\n";
-                foreach (var fight in boss)
-                {                    
-                    var fightDetails = fight.GetDetails();
-                    var timestamp = $"{fightDetails.Time - videoStart:hh\\:mm\\:ss}";
-
-                    string line;
-                    if (fightDetails.Result == "Kill")
-                        line = $"Kill - {timestamp}";
-                    else
-                    {
-                        i++;
-                        line = $"Wipe {i} - {timestamp}";
-                    }
-
-                    Console.WriteLine(line);
-                    description += $"{line}\n";
-                }
-
-                description += '\n';
-            }
-
-            TextCopy.ClipboardService.SetText(description);
-
-            Console.WriteLine("Press any key to close...");
+            //var events = parser.GetCombatLogEvents();
+            //Console.WriteLine($"{events.Count()} scanned.");
+            
+            sw.Stop();
+            Console.WriteLine($"Time taken: {sw.Elapsed}");
             Console.ReadKey();
-        }
-
-        private static void GetVideoTimestamps(IList<IFight> pulls, DateTime startOfVideo)
-        {
-
         }
     }
 }
